canal实时同步秒杀商品的静态页和ES索引

+++ xbd-vip-mall/mall-api/mall-seckill-api/src/main/java/com/xbd/vip/mall/seckill/model/SeckillGoods.java
@@ -28,7 +28,7 @@ public class SeckillGoods {
     private String content;
     private Integer price;
     @Column(name = "seckill_price")
-    private Integer seckillPrice; //seckill_price
+    private Double seckillPrice; //seckill_price
     private Integer num;
     @Column(name = "store_count")
     private Integer storeCount;
